% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spipLocpars}
\alias{spipLocpars}
\title{Write a locus information file for Spip}
\usage{
spipLocpars(genotypes, file, comment = "Generated using spipR")
}
\arguments{
\item{genotypes}{a data frame with columns working by pairs. Every
successive pair of columns represents the alleles for one locus.
Rows are individuals.}

\item{file}{file name used to save to locus information file}

\item{comment}{optional comment string}
}
\value{
A list with the parameters used
}
\description{
Write a locus information file for Spip. Most of the
  description come from \code{spip_man_page.pdf}. The format of this
  allele-frequency information file is simple. It includes only integers
  (or comments enclosed by a pair of & characters). The first integer is
  the number of loci in the file. Then for each locus you give the number
  of alleles at that locus followed by the allele frequencies of each
  allele at the locus. All these things should be sepa- rated by
  whitespace. No punctuation! If the allele frequencies at a locus do not
  sum to one they will be rescaled so that the do sum to one. If the number
  of allele frequencies listed does not match the number of alleles given
  for the locus then the program may fail without warning or give
  otherwise unexpected results.
  \code{NA} in the genotype input are not taken into account.
}
\examples{
n_ind = 10
locus1_1 = sample(1:5, n_ind, replace = T)
locus1_2 = sample(1:5, n_ind, replace = T)
locus2_1 = sample(1:8, n_ind, replace = T)
locus2_2 = sample(1:8, n_ind, replace = T)
locus3_1 = sample(c(1:3, NA), n_ind, replace = T)
locus3_2 = sample(1:3, n_ind, replace = T)
genotypes = data.frame(locus1_1, locus1_2,
  locus2_1, locus2_2,
  locus3_1, locus3_2)
a = spipLocpars(genotypes = genotypes, file = "toto.locpars")
}

